Adding a post by Glenn Gillen on Hiring.
His thoughts about interviews being more about making sure that both the
company and the candidate get as much information as possible to figure out
if they are suitable for one another really resonates with me. I also
really liked the idea of restoring the balance and respecting the time
of the candidates.